Refactors dev-env to split responsibilities
