+++ __tests__/lib/dev-env/dev-env-cli.js
@@ -1,0 +1,48 @@
+/**
+ * @format
+ */
+
+/**
+ * External dependencies
+ */
+
+/**
+ * Internal dependencies
+ */
+import { getEnvironmentName } from 'lib/dev-env/dev-env-cli';
+
+describe( 'lib/dev-env/dev-env-cli', () => {
+	describe( 'getEnvironmentName', () => {
+		it.each( [
+			{ // default value
+				options: {},
+				expected: 'vip-local',
+			},
+			{ // use custom name
+				options: {
+					slug: 'foo',
+				},
+				expected: 'foo',
+			},
+			{ // construct name from app and env
+				options: {
+					app: '123',
+					env: 'bar.car',
+				},
+				expected: '123-bar.car',
+			},
+			{ // custom name takes precedence
+				options: {
+					slug: 'foo',
+					app: '123',
+					env: 'bar.car',
+				},
+				expected: 'foo',
+			},
+		] )( 'should get correct name', async input => {
+			const result = getEnvironmentName( input.options );
+
+			expect( result ).toStrictEqual( input.expected );
+		} );
+	} );
+} );
