+++ __tests__/lib/dev-env/dev-env-core.js
@@ -18,12 +18,12 @@ import { getEnvironmentPath,
 	generateInstanceData,
 	destroyEnvironment,
 	getParamInstanceData,
-	getEnvironmentName } from 'lib/dev-environment';
+	getEnvironmentName } from 'lib/dev-env/dev-env-core';
 
 jest.mock( 'xdg-basedir', () => ( {} ) );
 jest.mock( 'fs' );
 
-describe( 'lib/dev-environment', () => {
+describe( 'lib/dev-env/dev-env-core', () => {
 	describe( 'createEnvironment', () => {
 		it( 'should throw for existing folder', async () => {
 			const slug = 'foo';
@@ -265,39 +265,6 @@ describe( 'lib/dev-environment', () => {
 			const filePath = getEnvironmentPath( name );
 
 			expect( filePath ).toBe( `${ os.tmpdir() }/vip/dev-environment/${ name }` );
-		} );
-	} );
-	describe( 'getEnvironmentName', () => {
-		it.each( [
-			{ // default value
-				options: {},
-				expected: 'vip-local',
-			},
-			{ // use custom name
-				options: {
-					slug: 'foo',
-				},
-				expected: 'foo',
-			},
-			{ // construct name from app and env
-				options: {
-					app: '123',
-					env: 'bar.car',
-				},
-				expected: '123-bar.car',
-			},
-			{ // custom name takes precedence
-				options: {
-					slug: 'foo',
-					app: '123',
-					env: 'bar.car',
-				},
-				expected: 'foo',
-			},
-		] )( 'should get correct name', async input => {
-			const result = getEnvironmentName( input.options );
-
-			expect( result ).toStrictEqual( input.expected );
 		} );
 	} );
 } );
