+++ src/bin/vip-dev-environment-info.js
@@ -14,7 +14,8 @@ import debugLib from 'debug';
  * Internal dependencies
  */
 import command from 'lib/cli/command';
-import { getEnvironmentName, printEnvironmentInfo, printAllEnvironmentsInfo, handleCLIException } from 'lib/dev-environment';
+import { printEnvironmentInfo, printAllEnvironmentsInfo } from 'lib/dev-env/dev-env-core';
+import { getEnvironmentName, handleCLIException } from 'lib/dev-env/dev-env-cli';
 import { DEV_ENVIRONMENT_FULL_COMMAND, DEV_ENVIRONMENT_SUBCOMMAND } from 'lib/constants/dev-environment';
 
 const debug = debugLib( '@automattic/vip:bin:dev-environment' );
