+++ src/bin/vip-dev-environment-start.js
@@ -14,7 +14,8 @@ import debugLib from 'debug';
  * Internal dependencies
  */
 import command from 'lib/cli/command';
-import { getEnvironmentName, startEnvironment, handleCLIException } from 'lib/dev-environment';
+import { startEnvironment } from 'lib/dev-env/dev-env-core';
+import { getEnvironmentName, handleCLIException } from 'lib/dev-env/dev-env-cli';
 import { DEV_ENVIRONMENT_FULL_COMMAND } from 'lib/constants/dev-environment';
 
 const debug = debugLib( '@automattic/vip:bin:dev-environment' );
