+++ src/bin/vip-dev-environment-wp.js
@@ -14,7 +14,8 @@ import debugLib from 'debug';
  * Internal dependencies
  */
 import command from 'lib/cli/command';
-import { getEnvironmentName, handleCLIException, runWp } from 'lib/dev-environment';
+import { getEnvironmentName, handleCLIException } from 'lib/dev-env/dev-env-cli';
+import { runWp } from 'lib/dev-env/dev-env-core';
 import { DEV_ENVIRONMENT_FULL_COMMAND } from 'lib/constants/dev-environment';
 
 const debug = debugLib( '@automattic/vip:bin:dev-environment' );
