+++ src/lib/dev-env/dev-env-cli.js
@@ -1,0 +1,55 @@
+/**
+ * @flow
+ * @format
+ */
+
+/**
+ * External dependencies
+ */
+import chalk from 'chalk';
+import formatters from 'lando/lib/formatters';
+
+/**
+ * Internal dependencies
+ */
+import { DEV_ENVIRONMENT_FULL_COMMAND } from '../constants/dev-environment';
+
+const DEFAULT_SLUG = 'vip-local';
+
+export function handleCLIException( exception: Error ) {
+	const errorPrefix = chalk.red( 'Error:' );
+	if ( 'Environment not found.' === exception.message ) {
+		const createCommand = chalk.bold( DEV_ENVIRONMENT_FULL_COMMAND + ' create' );
+
+		const message = `Environment doesn't exist.\n\n\nTo create a new environment run:\n\n${ createCommand }\n`;
+		console.log( errorPrefix, message );
+	} else {
+		console.log( errorPrefix, exception.message );
+	}
+}
+
+type EnvironmentNameOptions = {
+	slug: string,
+	app: string,
+	env: string,
+}
+
+export function getEnvironmentName( options: EnvironmentNameOptions ) {
+	if ( options.slug ) {
+		return options.slug;
+	}
+
+	if ( options.app ) {
+		const envSuffix = options.env ? `-${ options.env }` : '';
+
+		return options.app + envSuffix;
+	}
+
+	return DEFAULT_SLUG;
+}
+
+export function printTable( data: Object ) {
+	const formattedData = formatters.formatData( data, { format: 'table' }, { border: false } );
+
+	console.log( formattedData );
+}
