+++ src/lib/dev-env/dev-env-core.js
@@ -12,15 +12,13 @@ import os from 'os';
 import fs from 'fs';
 import ejs from 'ejs';
 import path from 'path';
-import Lando from 'lando/lib/lando';
-import landoUtils from 'lando/plugins/lando-core/lib/utils';
-import landoBuildTask from 'lando/plugins/lando-tooling/lib/build';
 import chalk from 'chalk';
 
 /**
  * Internal dependencies
  */
-import { DEV_ENVIRONMENT_FULL_COMMAND } from './constants/dev-environment';
+import { landoDestroy, landoInfo, landoRunWp, landoStart, landoStop } from './dev-env-lando';
+import { printTable } from './dev-env-cli';
 
 const debug = debugLib( '@automattic/vip:bin:dev-environment' );
 
@@ -43,7 +41,6 @@ const containerImages = {
 };
 
 export const defaults = {
-	environmentSlug: 'vip-local',
 	title: 'VIP Dev',
 	multisite: false,
 	phpVersion: '7.4',
@@ -63,8 +60,8 @@ export const defaults = {
 	};
 } );
 
-const landoFileTemplatePath = path.join( __dirname, '..', '..', 'assets', 'dev-environment.lando.template.yml.ejs' );
-const configDefaultsFilePath = path.join( __dirname, '..', '..', 'assets', 'dev-environment.wp-config-defaults.php' );
+const landoFileTemplatePath = path.join( __dirname, '..', '..', '..', 'assets', 'dev-environment.lando.template.yml.ejs' );
+const configDefaultsFilePath = path.join( __dirname, '..', '..', '..', 'assets', 'dev-environment.wp-config-defaults.php' );
 const landoFileName = '.lando.yml';
 
 export async function startEnvironment( slug: string ) {
@@ -174,11 +171,10 @@ export async function printEnvironmentInfo( slug: string ) {
 
 	const appInfo = await landoInfo( instancePath );
 
-	const lando = new Lando( getLandoConfig() );
-	console.log( lando.cli.formatData( appInfo, { format: 'table' }, { border: false } ) );
+	printTable( appInfo );
 }
 
-export async function runWp( slug: string, args: Array ) {
+export async function runWp( slug: string, args: Array<string> ) {
 	debug( 'Will run a wp command on env', slug, 'with args', args );
 
 	const instancePath = getEnvironmentPath( slug );
@@ -194,93 +190,6 @@ export async function runWp( slug: string, args: Array ) {
 	await landoRunWp( instancePath, args );
 }
 
-function getLandoConfig() {
-	const landoPath = path.join( __dirname, '..', '..', 'node_modules', 'lando' );
-
-	debug( `Getting lando config, using path '${ landoPath }' for plugins` );
-
-	return {
-		logLevelConsole: 'warn',
-		landoFile: '.lando.yml',
-		preLandoFiles: [ '.lando.base.yml', '.lando.dist.yml', '.lando.upstream.yml' ],
-		postLandoFiles: [ '.lando.local.yml' ],
-		pluginDirs: [
-			landoPath,
-			{
-				path: path.join( landoPath, 'integrations' ),
-				subdir: '.',
-			},
-		],
-	};
-}
-
-async function landoStart( instancePath ) {
-	debug( 'Will start lando app on path:', instancePath );
-
-	const lando = new Lando( getLandoConfig() );
-	await lando.bootstrap();
-
-	const app = lando.getApp( instancePath );
-	await app.init();
-
-	await app.start();
-
-	console.log( lando.cli.formatData( landoUtils.startTable( app ), { format: 'table' }, { border: false } ) );
-}
-
-async function landoStop( instancePath ) {
-	debug( 'Will stop lando app on path:', instancePath );
-
-	const lando = new Lando( getLandoConfig() );
-	await lando.bootstrap();
-
-	const app = lando.getApp( instancePath );
-	await app.init();
-
-	await app.stop();
-}
-
-async function landoDestroy( instancePath ) {
-	debug( 'Will destroy lando app on path:', instancePath );
-	const lando = new Lando( getLandoConfig() );
-	await lando.bootstrap();
-
-	const app = lando.getApp( instancePath );
-	await app.init();
-
-	await app.destroy();
-}
-
-async function landoInfo( instancePath ) {
-	const lando = new Lando( getLandoConfig() );
-	await lando.bootstrap();
-
-	const app = lando.getApp( instancePath );
-	await app.init();
-
-	const appInfo = landoUtils.startTable( app );
-
-	const reachableServices = app.info.filter( service => service.urls.length );
-	reachableServices.forEach( service => appInfo[ `${ service.service } urls` ] = service.urls );
-
-	const isUp = await isEnvUp( app );
-	appInfo.status = isUp ? chalk.green( 'UP' ) : chalk.yellow( 'DOWN' );
-
-	// Drop vipdev prefix
-	appInfo.name = appInfo.name.replace( /^vipdev/, '' );
-
-	return appInfo;
-}
-
-async function isEnvUp( app ) {
-	const reachableServices = app.info.filter( service => service.urls.length );
-	const urls = reachableServices.map( service => service.urls ).flat();
-
-	const scanResult = await app.scanUrls( urls, { max: 1 } );
-	// If all the URLs are reachable than the app is considered 'up'
-	return scanResult?.length && scanResult.filter( result => result.status ).length === scanResult.length;
-}
-
 async function prepareLandoEnv( instanceData, instancePath ) {
 	const landoFile = await ejs.renderFile( landoFileTemplatePath, instanceData );
 
@@ -294,44 +203,6 @@ async function prepareLandoEnv( instanceData, instancePath ) {
 	fs.copyFileSync( configDefaultsFilePath, configDefaultsFileTargetPath );
 
 	debug( `Lando file created in ${ landoFileTargetPath }` );
-}
-
-async function landoRunWp( instancePath, args ) {
-	const lando = new Lando( getLandoConfig() );
-	await lando.bootstrap();
-
-	const app = lando.getApp( instancePath );
-	await app.init();
-
-	const isUp = await isEnvUp( app );
-
-	if ( ! isUp ) {
-		throw new Error( 'environment needs to be started before running wp command' );
-	}
-
-	const wpTooling = app.config.tooling?.wp;
-	if ( ! wpTooling ) {
-		throw new Error( 'wp is not a known lando task' );
-	}
-
-	/*
-	 lando is looking in both passed args and process.argv so we need to do a bit of hack to fake process.argv
-	 so that lando doesn't try to interpret args not meant for wp.
-
-	 Lando drops first 3 args (<node> <lando> <command>) from process.argv and process rest, so we will fake 3 args + the real args
-	*/
-	process.argv = [ '0', '1', '3' ].concat( args );
-
-	wpTooling.app = app;
-	wpTooling.name = 'wp';
-
-	const wpTask = landoBuildTask( wpTooling, lando );
-
-	const argv = {
-		_: args,
-	};
-
-	wpTask.run( argv );
 }
 
 export function generateInstanceData( slug: string, options: NewInstanceOptions ) {
@@ -406,34 +277,3 @@ export function getEnvironmentPath( name: string ) {
 	return path.join( mainEnvironmentPath, 'vip', 'dev-environment', name );
 }
 
-export function handleCLIException( exception: Error ) {
-	const errorPrefix = chalk.red( 'Error:' );
-	if ( 'Environment not found.' === exception.message ) {
-		const createCommand = chalk.bold( DEV_ENVIRONMENT_FULL_COMMAND + ' create' );
-
-		const message = `Environment doesn't exist.\n\n\nTo create a new environment run:\n\n${ createCommand }\n`;
-		console.log( errorPrefix, message );
-	} else {
-		console.log( errorPrefix, exception.message );
-	}
-}
-
-type EnvironmentNameOptions = {
-	slug: string,
-	app: string,
-	env: string,
-}
-
-export function getEnvironmentName( options: EnvironmentNameOptions ) {
-	if ( options.slug ) {
-		return options.slug;
-	}
-
-	if ( options.app ) {
-		const envSuffix = options.env ? `-${ options.env }` : '';
-
-		return options.app + envSuffix;
-	}
-
-	return defaults.environmentSlug;
-}
